LayoutTests/svg/custom/svg-fonts-in-text-controls-expected.txt

This tests that the width of textareas and inputs is correctly calculated based on the metrics of the SVG font.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 141 is 141
PASS 140 is 140
PASS successfullyParsed is true

TEST COMPLETE
